Refactor kernel launching & event recording.
No semantic change intended in kernel launches,
OTOH hip events have been changed to use cl_events
from actual commands instead of Markers. This has the
advantage that the numbers (from hipEventElapsedTime)
are much more closer to reality then previously.

The difficulty is that OpenCL has no concept of
"event recording", instead events are returned by
each enqueued command. The simple approach (taken
here) is to always retain the last enqueued command's
cl_event, and attach that to the hipEvent_t when
the user calls hipRecordEvent(). Downside is that
if some time passes between the last event's finish
and the actual call of hipRecordEvent, it will be
included in the result of hipEventElapsedTime().
This should only happen though, if there is a delay
between the end of last event and the next call of hipRecordEvent.